+++ src/Equinor.ProCoSys.Completion.Command/Validators/LibraryItemValidator.cs
@@ -7,7 +7,7 @@ using Equinor.ProCoSys.Completion.Domain.AggregateModels.LibraryAggregate;
 using Equinor.ProCoSys.Completion.Domain.Validators;
 using Microsoft.EntityFrameworkCore;
 
-namespace Equinor.ProCoSys.Completion.WebApi.Validators;
+namespace Equinor.ProCoSys.Completion.Command.Validators;
 
 public class LibraryItemValidator : ILibraryItemValidator
 {
