+++ src/Equinor.ProCoSys.Completion.Command/Validators/PersonValidator.cs
@@ -8,7 +8,7 @@ using Equinor.ProCoSys.Completion.Domain.AggregateModels.PersonAggregate;
 using Equinor.ProCoSys.Completion.Domain.Validators;
 using Microsoft.EntityFrameworkCore;
 
-namespace Equinor.ProCoSys.Completion.WebApi.Validators;
+namespace Equinor.ProCoSys.Completion.Command.Validators;
 
 public class PersonValidator : IPersonValidator
 {
