+++ src/Equinor.ProCoSys.Completion.WebApi/DiModules/ApplicationModule.cs
@@ -28,7 +28,7 @@ using Equinor.ProCoSys.Completion.WebApi.Authorizations;
 using Equinor.ProCoSys.Completion.WebApi.Controllers;
 using Equinor.ProCoSys.Completion.WebApi.MassTransit;
 using Equinor.ProCoSys.Completion.WebApi.Misc;
-using Equinor.ProCoSys.Completion.WebApi.Validators;
+using Equinor.ProCoSys.Completion.Command.Validators;
 using MassTransit;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -125,9 +125,9 @@ public static class ApplicationModule
         services.AddScoped<IWorkOrderValidator, WorkOrderValidator>();
         services.AddScoped<ISWCRValidator, SWCRValidator>();
         services.AddScoped<IDocumentValidator, DocumentValidator>();
-        services.AddScoped<ICheckListValidator, CheckListValidator>();
-        services.AddScoped<IRowVersionValidator, RowVersionValidator>();
-        services.AddScoped<IPatchOperationValidator, PatchOperationValidator>();
+        services.AddScoped<ICheckListValidator, ProCoSys4CheckListValidator>();
+        services.AddScoped<IRowVersionInputValidator, RowVersionInputValidator>();
+        services.AddScoped<IPatchOperationInputValidator, PatchOperationInputValidator>();
 
         services.AddScoped<IAzureBlobService, AzureBlobService>();
