+++ src/tests/Equinor.ProCoSys.Completion.Command.Tests/Validators/DocumentValidatorTests.cs
@@ -3,11 +3,11 @@ using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.Domain.AggregateModels.DocumentAggregate;
 using Equinor.ProCoSys.Completion.Infrastructure;
 using Equinor.ProCoSys.Completion.Test.Common;
-using Equinor.ProCoSys.Completion.WebApi.Validators;
+using Equinor.ProCoSys.Completion.Command.Validators;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Equinor.ProCoSys.Completion.WebApi.Tests.Validators;
+namespace Equinor.ProCoSys.Completion.Command.Tests.Validators;
 
 [TestClass]
 public class DocumentValidatorTests : ReadOnlyTestsBase
