+++ src/tests/Equinor.ProCoSys.Completion.Command.Tests/Validators/ProjectValidatorTests.cs
@@ -1,54 +1,67 @@
 ﻿using System;
 using System.Threading.Tasks;
-using Equinor.ProCoSys.Completion.Domain.AggregateModels.WorkOrderAggregate;
+using Equinor.ProCoSys.Completion.Domain.AggregateModels.ProjectAggregate;
 using Equinor.ProCoSys.Completion.Infrastructure;
 using Equinor.ProCoSys.Completion.Test.Common;
-using Equinor.ProCoSys.Completion.WebApi.Validators;
+using Equinor.ProCoSys.Completion.Command.Validators;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Equinor.ProCoSys.Completion.WebApi.Tests.Validators;
+namespace Equinor.ProCoSys.Completion.Command.Tests.Validators;
 
 [TestClass]
-public class WorkOrderValidatorTests : ReadOnlyTestsBase
+public class ProjectValidatorTests : ReadOnlyTestsBase
 {
-    private WorkOrder _knownWorkOrder = null!;
-    private WorkOrder _openWorkOrder = null!;
-    private WorkOrder _closedWorkOrder = null!;
+    private Project _openProject = null!;
+    private Project _closedProject = null!;
 
     protected override void SetupNewDatabase(DbContextOptions<CompletionContext> dbContextOptions)
     {
         using var context = new CompletionContext(dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
-
-        _knownWorkOrder = _openWorkOrder = new WorkOrder(TestPlantA, Guid.NewGuid(), "WorkOrder 1");
-        _closedWorkOrder = new WorkOrder(TestPlantA, Guid.NewGuid(), "WorkOrder 2") { IsClosed = true };
-        context.WorkOrders.Add(_openWorkOrder);
-        context.WorkOrders.Add(_closedWorkOrder);
+            
+        _openProject = new Project(TestPlantA, Guid.NewGuid(), "Project 1", "D1");
+        _closedProject = new Project(TestPlantA, Guid.NewGuid(), "Project 2", "D2") { IsClosed = true };
+        context.Projects.Add(_openProject);
+        context.Projects.Add(_closedProject);
 
         context.SaveChangesAsync().Wait();
     }
 
     #region ExistsAsync
     [TestMethod]
-    public async Task ExistsAsync_ShouldReturnTrue_WhenWorkOrderExist()
+    public async Task ExistsAsync_ShouldReturnTrue_WhenProjectIsClosed()
     {
         // Arrange
-        await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
-        var dut = new WorkOrderValidator(context);
+        await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);            
+        var dut = new ProjectValidator(context);
 
         // Act
-        var result = await dut.ExistsAsync(_knownWorkOrder.Guid, default);
+        var result = await dut.ExistsAsync(_closedProject.Guid, default);
 
         // Assert
         Assert.IsTrue(result);
     }
 
     [TestMethod]
-    public async Task ExistsAsync_ShouldReturnFalse_WhenWorkOrderNotExist()
+    public async Task ExistsAsync_ShouldReturnTrue_WhenProjectIsOpen()
+    {
+        // Arrange
+        await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
+        var dut = new ProjectValidator(context);
+
+        // Act
+        var result = await dut.ExistsAsync(_openProject.Guid, default);
+
+        // Assert
+        Assert.IsTrue(result);
+    }
+
+    [TestMethod]
+    public async Task ExistsAsync_ShouldReturnFalse_WhenProjectNotExist()
     {
         // Arrange
         await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);    
-        var dut = new WorkOrderValidator(context);
+        var dut = new ProjectValidator(context);
 
         // Act
         var result = await dut.ExistsAsync(Guid.Empty, default);
@@ -60,39 +73,39 @@ public class WorkOrderValidatorTests : ReadOnlyTestsBase
 
     #region IsClosed
     [TestMethod]
-    public async Task IsClosed_ShouldReturnTrue_WhenWorkOrderIsClosed()
+    public async Task IsClosed_ShouldReturnTrue_WhenProjectIsClosed()
     {
         // Arrange
         await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
-        var dut = new WorkOrderValidator(context);
+        var dut = new ProjectValidator(context);
 
         // Act
-        var result = await dut.IsClosedAsync(_closedWorkOrder.Guid, default);
+        var result = await dut.IsClosedAsync(_closedProject.Guid, default);
 
         // Assert
         Assert.IsTrue(result);
     }
 
     [TestMethod]
-    public async Task IsClosed_ShouldReturnFalse_WhenWorkOrderIsOpen()
+    public async Task IsClosed_ShouldReturnFalse_WhenProjectIsOpen()
     {
         // Arrange
         await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
-        var dut = new WorkOrderValidator(context);
+        var dut = new ProjectValidator(context);
 
         // Act
-        var result = await dut.IsClosedAsync(_openWorkOrder.Guid, default);
+        var result = await dut.IsClosedAsync(_openProject.Guid, default);
 
         // Assert
         Assert.IsFalse(result);
     }
 
     [TestMethod]
-    public async Task IsClosed_ShouldReturnFalse_WhenWorkOrderNotExist()
+    public async Task IsClosed_ShouldReturnFalse_WhenProjectNotExist()
     {
         // Arrange
         await using var context = new CompletionContext(_dbContextOptions, _plantProviderMockObject, _eventDispatcherMockObject, _currentUserProviderMockObject);
-        var dut = new WorkOrderValidator(context);
+        var dut = new ProjectValidator(context);
 
         // Act
         var result = await dut.IsClosedAsync(Guid.Empty, default);
