+++ src/tests/Equinor.ProCoSys.Completion.WebApi.Tests/Controllers/Attachments/OverwriteAttachmentDtoValidatorTests.cs
@@ -10,11 +10,11 @@ namespace Equinor.ProCoSys.Completion.WebApi.Tests.Controllers.Attachments;
 public class OverwriteAttachmentDtoValidatorTests : UploadBaseDtoValidatorTests<OverwriteAttachmentDto>
 {
     private readonly string _rowVersion = "AAAAAAAAABA=";
-    private IRowVersionValidator _rowVersionValidatorMock = null!;
+    private IRowVersionInputValidator _rowVersionValidatorMock = null!;
 
     protected override void SetupDut()
     {
-        _rowVersionValidatorMock = Substitute.For<IRowVersionValidator>();
+        _rowVersionValidatorMock = Substitute.For<IRowVersionInputValidator>();
         _rowVersionValidatorMock.IsValid(_rowVersion)
             .Returns(true);
         _dut = new OverwriteAttachmentDtoValidator(
