+++ src/tests/Equinor.ProCoSys.Completion.WebApi.Tests/Controllers/RowVersionDtoValidatorTests.cs
@@ -11,12 +11,12 @@ public class RowVersionDtoValidatorTests
     private readonly string _rowVersion = "AAAAAAAAABA=";
 
     private RowVersionDtoValidator _dut = null!;
-    private IRowVersionValidator _rowVersionValidatorMock = null!;
+    private IRowVersionInputValidator _rowVersionValidatorMock = null!;
 
     [TestInitialize]
     public void Setup_OkState()
     {
-        _rowVersionValidatorMock = Substitute.For<IRowVersionValidator>();
+        _rowVersionValidatorMock = Substitute.For<IRowVersionInputValidator>();
         _rowVersionValidatorMock.IsValid(_rowVersion).Returns(true);
 
         _dut = new RowVersionDtoValidator(_rowVersionValidatorMock);
